requirements.txt
# Django Core
Django==5.2.4
django-jazzmin==3.0.1
django-widget-tweaks==1.5.0
dj-database-url==3.0.1

# Database & ORM
sqlparse==0.5.3

# Authentication & User Management
# (Django built-in auth is used)

# Static Files & Media
whitenoise==6.9.0
Pillow==11.3.0

# PDF Generation
playwright==1.48.0
weasyprint==66.1
pdf2image==1.17.0

# OCR & Image Processing
pytesseract==0.3.13
opencv-python-headless==4.12.0.88
easyocr==1.7.2

# AI & Machine Learning
openai==1.99.1
transformers==4.55.2
torch==2.8.0
torchvision==0.23.0
datasets==4.0.0
scikit-image==0.25.2
scipy==1.16.1
numpy==2.2.6
pandas==2.3.1

# Text Processing & NLP
nltk==3.9.1
langdetect==1.0.9
jieba3k==0.35.1
tinysegmenter==0.3
sentencepiece==0.2.1
tokenizers==0.21.4
regex==2025.7.34

# Text-to-Speech
gTTS==2.5.4
elevenlabs==2.9.2

# Web Scraping & RSS
beautifulsoup4==4.13.4
lxml==6.0.0
feedparser==6.0.11
newspaper3k==0.2.8
requests==2.32.4
aiohttp==3.12.15
httpx==0.28.1

# Data Processing
pyarrow==21.0.0
duckdb==1.3.2
joblib==1.5.1
multiprocess==0.70.16

# Utilities
python-dotenv==1.1.1
python-dateutil==2.9.0.post0
pytz==2025.2
tzdata==2025.2
PyYAML==6.0.2
click==8.1.8
tqdm==4.67.1

# Development & Testing
# (Add these for development environment)
# pytest==8.0.0
# pytest-django==4.8.0
# coverage==7.4.0

# Production
# (Add these for production deployment)
# gunicorn==23.0.0
# psycopg2-binary==2.9.9  # For PostgreSQL
# redis==5.0.1  # For caching
# celery==5.3.4  # For background tasks

# Note: Some packages may have system dependencies
# - pytesseract requires tesseract-ocr to be installed on the system
# - opencv-python-headless may require system libraries
# - weasyprint may require system fonts and libraries
